# docs/swagger.yaml
basePath: /api
definitions:
  models.Airport:
    properties:
      eventId:
        type: integer
      id:
        type: integer
      identifier:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      maximumAircraftPerHour:
        type: integer
      numberOfVotes:
        type: integer
    type: object
...
  models.DepartureTimeWindowOffsetsCalculationMode:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - DepartureTimeWindowOffsetsCalculationModeNone
    - DepartureTimeWindowOffsetsCalculationModeEarliestRoutes
    - DepartureTimeWindowOffsetsCalculationModeLatestRoutes
    - DepartureTimeWindowOffsetsCalculationModeRouteAverage
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  models.HighlightedWaypoint:
    properties:
      color:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      identifier:
        type: string
      note:
        type: string
    type: object
  models.Location:
    properties:
      eventId:
        type: integer
      id:
        type: integer
      identifier:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      maximumAircraftPerHour:
        type: integer
      routeSegmentId:
        type: integer
      sortOrder:
        type: integer
    type: object
...
  models.RouteSegment:
    properties:
      color:
        type: string
      enabled:
        type: boolean
      eventId:
        type: integer
      facilities:
        type: string
      id:
        type: integer
      identifier:
        type: string
      locations:
        items:
          $ref: '#/definitions/models.Location'
        type: array
      maximumAircraftPerHour:
        type: integer
      providedFacilityProgression:
        items:
          $ref: '#/definitions/models.Sector'
        type: array
      routeRevision:
        type: integer
      routeSegmentGroup:
        type: string
      routeString:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.RouteSegmentTag'
        type: array
    type: object
  models.RouteSegmentTag:
    properties:
      id:
        type: integer
      routeSegmentId:
        type: integer
      tag:
        type: string
    type: object
  models.Sector:
    properties:
      eventId:
        type: integer
      id:
        type: integer
      identifier:
        type: string
      maximumAircraftPerHour:
        type: integer
      sectorBoundaries:
        items:
          $ref: '#/definitions/models.SectorBoundary'
        type: array
    type: object
  models.SectorBoundary:
    properties:
      coordinates:
        items:
          $ref: '#/definitions/models.SectorBoundaryCoordinate'
        type: array
      id:
        type: integer
      maxLatitude:
        type: number
      maxLongitude:
        type: number
      minLatitude:
        type: number
      minLongitude:
        type: number
      sectorId:
        type: integer
    type: object
  models.SectorBoundaryCoordinate:
    properties:
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      sectorBoundaryId:
        type: integer
      sortOrder:
        type: integer
    type: object
  models.SimulatorDataResponse:
    properties:
      event:
        $ref: '#/definitions/models.VATSIMEvent'
      revision:
        $ref: '#/definitions/models.SlotRevision'
    type: object
  models.Slot:
    properties:
      arrivalAirport:
        $ref: '#/definitions/models.Airport'
      arrivalAirportId:
        type: integer
      departureAirport:
        $ref: '#/definitions/models.Airport'
      departureAirportId:
        type: integer
      departureTime:
        type: string
      id:
        type: integer
      projectedArrivalTime:
        type: string
      routeSegments:
        items:
          $ref: '#/definitions/models.RouteSegment'
        type: array
      slotRevisionId:
        type: integer
    type: object
  models.SlotGenerationMode:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - SlotGenerationModeMaximizeSlots
    - SlotGenerationModeRandom
  models.SlotRevision:
    properties:
      createdAt:
        type: string
      eventId:
        type: integer
      id:
        type: integer
      number:
        type: integer
      simulationOutputCommentary:
        type: string
      slotGenerationOutputCommentary:
        type: string
      slots:
        items:
          $ref: '#/definitions/models.Slot'
        type: array
      throughputSnapshots:
        items:
          $ref: '#/definitions/models.ThroughputSnapshot'
        type: array
      throughputStates:
        items:
          $ref: '#/definitions/models.ThroughputState'
        type: array
    type: object
  models.SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
  models.ThroughputSnapshot:
    properties:
      id:
        type: integer
      minuteOffset:
        type: integer
      slot:
        $ref: '#/definitions/models.Slot'
      slotId:
        type: integer
      slotRevisionId:
        type: integer
      throughputPointId:
        type: integer
      throughputPointType:
        type: string
    type: object
  models.ThroughputState:
    properties:
      departureTimeWindowStart:
        type: string
      id:
        type: integer
      maximumSlots:
        type: integer
      slotRevisionId:
        type: integer
      slotsAllocated:
        type: integer
      throughputPointId:
        type: integer
      throughputPointType:
        type: string
    type: object
  models.VATSIMEvent:
    properties:
      airports:
        items:
          $ref: '#/definitions/models.Airport'
        type: array
      calculateThroughputDataOnlyForManuallyProvidedSectors:
        type: boolean
      calculationFallbackGroundSpeed:
        type: number
      date:
        type: string
      departureTimeWindow:
        example: 3h0m0s
        type: string
      departureTimeWindowOffsetSynchronizationLongitude:
        type: number
      departureTimeWindowOffsetSynchronizationTimeOfDay:
        type: string
      highSimulationAccuracy:
        type: boolean
      id:
        type: integer
      intendedDepartureTimeWindowOffsetsCalculationMode:
        $ref: '#/definitions/models.DepartureTimeWindowOffsetsCalculationMode'
      intendedSlotGenerationMode:
        $ref: '#/definitions/models.SlotGenerationMode'
      intendedWaypointThroughputCalculationMode:
        $ref: '#/definitions/models.WaypointThroughputCalculationMode'
      recalculateMaximumAirportSlots:
        type: boolean
      routeRevision:
        type: integer
      routeSegments:
        items:
          $ref: '#/definitions/models.RouteSegment'
        type: array
      sectors:
        items:
          $ref: '#/definitions/models.Sector'
        type: array
      shouldSimulationUseActualWeatherForecastData:
        type: boolean
      simulationAnalysisResolutionInMinutes:
        type: integer
      slotRevisions:
        items:
          $ref: '#/definitions/models.SlotRevision'
        type: array
      thresholdToCheckIfAirplaneIsCountedAtWaypointInNm:
        type: number
      title:
        type: string
      waypoints:
        items:
          $ref: '#/definitions/models.Location'
        type: array
    type: object
  models.WaypointThroughputCalculationMode:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - WaypointThroughputCalculationModeNone
    - WaypointThroughputCalculationModeFirstWaypointsOfNATRouteSegmentsOnly
    - WaypointThroughputCalculationModeAllWaypoints
host: localhost:3000
info:
  contact: {}
  description: Central data API for CTP (Cross the Pond) event management. All /api/*
    endpoints require an X-API-Key header.
  title: CTP API
  version: 1.0.0
paths:
...
  /events:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.VATSIMEvent'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all events
      tags:
      - events
    post:
      consumes:
      - application/json
      parameters:
      - description: Event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.VATSIMEvent'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.VATSIMEvent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create event
      tags:
      - events
...
  /events/{id}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete event
      tags:
      - events
    get:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VATSIMEvent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get event with all nested data
      tags:
      - events
    put:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Event fields to update
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.VATSIMEvent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VATSIMEvent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update event
      tags:
      - events
  /events/{id}/simulator-data:
    get:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Specific slot revision number (defaults to latest)
        in: query
        name: revision
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SimulatorDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get full simulator data for an event
      tags:
      - events
  /highlighted-waypoints:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.HighlightedWaypoint'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all highlighted waypoints
      tags:
      - highlighted-waypoints
    post:
      consumes:
      - application/json
      parameters:
      - description: Highlighted waypoint
        in: body
        name: waypoint
        required: true
        schema:
          $ref: '#/definitions/models.HighlightedWaypoint'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.HighlightedWaypoint'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create or update highlighted waypoint (upsert by identifier)
      tags:
      - highlighted-waypoints
  /highlighted-waypoints/{identifier}:
    delete:
      parameters:
      - description: Waypoint identifier
        in: path
        name: identifier
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete highlighted waypoint
      tags:
      - highlighted-waypoints
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"